refactor(model): Renamed 'entities' package to 'model'
