+++ src/main/java/edu/ntnu/idi/idatt/model/BoardGame.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idi.idatt.entities;
+package edu.ntnu.idi.idatt.model;
 
 import java.util.ArrayList;
 import java.util.HashMap;
